+++ src/main/java/com/nhnacademy/store99/bookstore/book_category/dto/response/BookCategoryResponse.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.store99.bookstore.book_category.dto.response;
+
+public class BookCategoryResponse {
+}
