+++ src/main/java/com/nhnacademy/store99/bookstore/book_category/repository/Impl/BookCategoryRepositoryImpl.java
@@ -2,16 +2,14 @@ package com.nhnacademy.store99.bookstore.book_category.repository.Impl;
 
 import com.nhnacademy.store99.bookstore.author.entity.QAuthor;
 import com.nhnacademy.store99.bookstore.book.entity.QBook;
-import com.nhnacademy.store99.bookstore.book.response.BookListElementDTO;
+import com.nhnacademy.store99.bookstore.book.dto.response.response.BookListElementDTO;
 import com.nhnacademy.store99.bookstore.book_author.entity.QBookAuthor;
 import com.nhnacademy.store99.bookstore.book_category.entity.BookCategory;
 import com.nhnacademy.store99.bookstore.book_category.entity.QBookCategory;
 import com.nhnacademy.store99.bookstore.book_category.repository.BookCategoryRepository;
-import com.nhnacademy.store99.bookstore.book_category.response.BookCategoryResponse;
-import com.nhnacademy.store99.bookstore.book_category.response.CategoryParentsDTO;
-import com.nhnacademy.store99.bookstore.book_tag.entity.QBookTag;
+import com.nhnacademy.store99.bookstore.book_category.dto.response.BookCategoryResponse;
+import com.nhnacademy.store99.bookstore.book_category.dto.response.CategoryParentsDTO;
 import com.nhnacademy.store99.bookstore.category.entity.QCategory;
-import com.nhnacademy.store99.bookstore.tag.entity.QTag;
 import com.querydsl.core.group.GroupBy;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.JPQLQuery;
@@ -70,9 +68,7 @@ public class BookCategoryRepositoryImpl extends QuerydslRepositorySupport implem
         QBookCategory bookCategory = QBookCategory.bookCategory;
         QBook book = QBook.book;
         QAuthor author = QAuthor.author;
-        QTag tag = QTag.tag;
         QBookAuthor bookAuthor = QBookAuthor.bookAuthor;
-        QBookTag bookTag = QBookTag.bookTag;
 
         List<Long> parentIds =
                 parentsDTOList.stream().map(CategoryParentsDTO::getCategoryId).collect(Collectors.toList());
@@ -113,23 +109,6 @@ public class BookCategoryRepositoryImpl extends QuerydslRepositorySupport implem
                         )
                 );
 
-        /**
-         * book id를 그룹화하여 tag id에 해당하는 tagName 담기
-         *
-         * @author rosin23
-         */
-
-        Map<Long, List<BookListElementDTO.TagDTO>> tagMap = from(bookTag)
-                .where(bookTag.book.id.in(bookIds))
-                .join(bookTag.tag, tag)
-                .transform(GroupBy.groupBy(bookTag.book.id)
-                        .as(GroupBy.list(
-                                        Projections.constructor(BookListElementDTO.TagDTO.class,
-                                                tag.tagName)
-                                )
-                        )
-                );
-
 
         List<BookListElementDTO> bookResponses = bookResponsesDtoVar.stream().map(b ->
                 {
@@ -146,7 +125,6 @@ public class BookCategoryRepositoryImpl extends QuerydslRepositorySupport implem
                             .BookStock(b.getBookStock())
                             .BookAvgOfRate(b.getBookAvgOfRate())
                             .authorsDTOList(authorsMap.get(b.getBookId()))
-                            .tagDTOList(tagMap.get(b.getBookId()))
                             .build();
                 }
         ).collect(Collectors.toList());
